split synch, Asynch and Note methods
