+++ app/src/main/java/com/edurda77/filmlibrary/data/RetrofitTheMdbRepoUseCaseImplAsync.kt
@@ -1,105 +1,29 @@
 package com.edurda77.filmlibrary.data
 
-import com.edurda77.filmlibrary.BuildConfig.TMDB_API_KEY
+import com.edurda77.filmlibrary.BuildConfig
 import com.edurda77.filmlibrary.data.retrofit.TheMDBRepoApi
-import com.edurda77.filmlibrary.domain.NoteRepo
-import com.edurda77.filmlibrary.domain.TheMDBRepoUseCace
+import com.edurda77.filmlibrary.domain.TheMDBRepoUseCaseAsync
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
+class RetrofitTheMdbRepoUseCaseImplAsync : TheMDBRepoUseCaseAsync{
 
-private const val BASE_URL = "https://api.themoviedb.org/3/"
-const val LANGUAGE = "ru-RU"
-
-
-class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
-
-
-    private val cacheNotes: MutableList<NoteMovie> = mutableListOf()
 
     private var retrofit: Retrofit = Retrofit.Builder()
         .baseUrl(BASE_URL)
         .addConverterFactory(GsonConverterFactory.create())
         .build()
     private var api: TheMDBRepoApi = retrofit.create(TheMDBRepoApi::class.java)
-
-
-    override fun getReposForSearchMovieSync(
-        userName: String,
-        adultKey: Boolean
-    ): List<ResultSearchMovie> {
-
-        val resultsParsing: ResultsParsing? =
-            api.getSearchMovie(TMDB_API_KEY, LANGUAGE, userName, adultKey)
-                .execute().body()
-        return parsingForSync(resultsParsing)
-
-    }
-
-
-    override fun getReposForIDMovieSync(searcheMovie: ResultSearchMovie): Movie? {
-
-        return api.getIDMovie(searcheMovie.id, TMDB_API_KEY, LANGUAGE)
-            .execute().body()
-    }
-
-
-    override fun getReposForGenresSync(): List<Genres> {
-
-        return api.getGenres(TMDB_API_KEY).execute().body() ?: emptyList()
-
-    }
-
-    override fun getReposForNowPlayingMovieSync(): List<ResultSearchMovie> {
-        val resultsParsing: ResultsParsing? = api.getNowPlaying(TMDB_API_KEY, LANGUAGE)
-            .execute().body()
-
-        return parsingForSync(resultsParsing)
-    }
-
-    override fun getReposForPopularMovieSync(): List<ResultSearchMovie> {
-        val resultsParsing: ResultsParsing? = api.getPopular(TMDB_API_KEY, LANGUAGE)
-            .execute().body()
-
-        return parsingForSync(resultsParsing)
-    }
-
-    override fun getReposForTopRatedMovieSync(): List<ResultSearchMovie> {
-        val resultsParsing: ResultsParsing? = api.getTopRated(TMDB_API_KEY, LANGUAGE)
-            .execute().body()
-
-        return parsingForSync(resultsParsing)
-    }
-
-    override fun getReposForUpcomingMovieSync(): List<ResultSearchMovie> {
-        val resultsParsing: ResultsParsing? = api.getUpcoming(TMDB_API_KEY, LANGUAGE)
-            .execute().body()
-
-        return parsingForSync(resultsParsing)
-    }
-
-    override fun getReposForSearchPeopleSync(name: String): List<ResultSearchedPeople> {
-        val resultsParsingPeople: ResultParsingPeople? =
-            api.getSearshPeople(TMDB_API_KEY, LANGUAGE, name)
-                .execute().body()
-        return parsingForPeopleSync(resultsParsingPeople)
-    }
-
-    override fun getReposForIdPeopleSync(searchedPeople: ResultSearchedPeople): People? {
-        return api.getIdPeople(searchedPeople.idPeopleSearched, TMDB_API_KEY, LANGUAGE)
-            .execute().body()
-    }
-
     override fun getReposForSearchMovieAsync(
         userName: String,
         adultKey: Boolean,
         onSuccess: (List<ResultSearchMovie>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getSearchMovie(TMDB_API_KEY, LANGUAGE, userName, adultKey)
+        api.getSearchMovie(BuildConfig.TMDB_API_KEY, LANGUAGE, userName, adultKey)
             .enqueue(object : Callback<ResultsParsing> {
                 override fun onResponse(
                     call: Call<ResultsParsing>,
@@ -129,7 +53,8 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (Movie) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getIDMovie(searcheMovie.id, TMDB_API_KEY, LANGUAGE).enqueue(object : Callback<Movie> {
+        api.getIDMovie(searcheMovie.id, BuildConfig.TMDB_API_KEY, LANGUAGE).enqueue(object :
+            Callback<Movie> {
             override fun onResponse(
                 call: Call<Movie>,
                 response: Response<Movie>
@@ -154,7 +79,7 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (List<Genres>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getGenres(TMDB_API_KEY).enqueue(object : Callback<List<Genres>> {
+        api.getGenres(BuildConfig.TMDB_API_KEY).enqueue(object : Callback<List<Genres>> {
             override fun onResponse(call: Call<List<Genres>>, response: Response<List<Genres>>) {
                 if (response.isSuccessful) {
                     onSuccess(response.body() ?: throw IllegalStateException("Нулевой результат"))
@@ -174,7 +99,8 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (List<ResultSearchMovie>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getNowPlaying(TMDB_API_KEY, LANGUAGE).enqueue(object : Callback<ResultsParsing> {
+        api.getNowPlaying(BuildConfig.TMDB_API_KEY, LANGUAGE).enqueue(object :
+            Callback<ResultsParsing> {
             override fun onResponse(
                 call: Call<ResultsParsing>,
                 response: Response<ResultsParsing>
@@ -201,7 +127,8 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (List<ResultSearchMovie>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getPopular(TMDB_API_KEY, LANGUAGE).enqueue(object : Callback<ResultsParsing> {
+        api.getPopular(BuildConfig.TMDB_API_KEY, LANGUAGE).enqueue(object :
+            Callback<ResultsParsing> {
             override fun onResponse(
                 call: Call<ResultsParsing>,
                 response: Response<ResultsParsing>
@@ -228,7 +155,8 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (List<ResultSearchMovie>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getTopRated(TMDB_API_KEY, LANGUAGE).enqueue(object : Callback<ResultsParsing> {
+        api.getTopRated(BuildConfig.TMDB_API_KEY, LANGUAGE).enqueue(object :
+            Callback<ResultsParsing> {
             override fun onResponse(
                 call: Call<ResultsParsing>,
                 response: Response<ResultsParsing>
@@ -255,7 +183,8 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (List<ResultSearchMovie>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getUpcoming(TMDB_API_KEY, LANGUAGE).enqueue(object : Callback<ResultsParsing> {
+        api.getUpcoming(BuildConfig.TMDB_API_KEY, LANGUAGE).enqueue(object :
+            Callback<ResultsParsing> {
             override fun onResponse(
                 call: Call<ResultsParsing>,
                 response: Response<ResultsParsing>
@@ -283,7 +212,7 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (List<ResultSearchedPeople>) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getSearshPeople(TMDB_API_KEY, LANGUAGE, name)
+        api.getSearshPeople(BuildConfig.TMDB_API_KEY, LANGUAGE, name)
             .enqueue(object : Callback<ResultParsingPeople> {
                 override fun onResponse(
                     call: Call<ResultParsingPeople>,
@@ -313,73 +242,25 @@ class RetrofitTheMdbRepoUseCaseImpl : TheMDBRepoUseCace, NoteRepo {
         onSuccess: (People) -> Unit,
         OnError: (Throwable) -> Unit
     ) {
-        api.getIdPeople(searchedPeople.idPeopleSearched, TMDB_API_KEY, LANGUAGE)
+        api.getIdPeople(searchedPeople.idPeopleSearched, BuildConfig.TMDB_API_KEY, LANGUAGE)
             .enqueue(object : Callback<People> {
-            override fun onResponse(
-                call: Call<People>,
-                response: Response<People>
-            ) {
-                if (response.isSuccessful) {
-                    onSuccess(response.body()!!)
-                } else {
-                    OnError(Throwable("Неизвестная ошибка"))
+                override fun onResponse(
+                    call: Call<People>,
+                    response: Response<People>
+                ) {
+                    if (response.isSuccessful) {
+                        onSuccess(response.body()!!)
+                    } else {
+                        OnError(Throwable("Неизвестная ошибка"))
+                    }
+
                 }
 
-            }
-
-            override fun onFailure(call: Call<People>, t: Throwable) {
-                OnError(t)
-            }
+                override fun onFailure(call: Call<People>, t: Throwable) {
+                    OnError(t)
+                }
 
 
-        })
-    }
-
-    private fun parsingForSync(resultsParsing: ResultsParsing?): List<ResultSearchMovie> {
-        val resultSearch = emptyList<ResultSearchMovie>().toMutableList()
-
-        resultsParsing?.results?.forEach {
-            resultSearch.add(it)
-        }
-        return resultSearch
-    }
-
-    private fun parsingForPeopleSync(resultsParsing: ResultParsingPeople?): List<ResultSearchedPeople> {
-        val resultSearch = emptyList<ResultSearchedPeople>().toMutableList()
-
-        resultsParsing?.results?.forEach {
-            resultSearch.add(it)
-        }
-        return resultSearch
-    }
-
-    override fun add(note: NoteMovie) {
-        cacheNotes.add(note)
-    }
-
-    override fun getNots(): List<NoteMovie> {
-        return ArrayList<NoteMovie>(cacheNotes)
-    }
-
-    override fun delete(id: Int) {
-        val indexOfDelete = cacheNotes.indexOfFirst {
-            it.idNote == id
-        }
-        cacheNotes.removeAt(indexOfDelete)
-
-    }
-
-    override fun update(id: Int, note: NoteMovie) {
-        val indexOfDelete = cacheNotes.indexOfFirst {
-            it.idNote == id
-        }
-        cacheNotes.removeAt(indexOfDelete)
-        cacheNotes.add(indexOfDelete, note)
-    }
-
-    override fun clearNots() {
-        cacheNotes.clear()
-
-
+            })
     }
 }
