--- app/src/main/java/com/edurda77/filmlibrary/domain/TheMDBRepoUseCac.kt
@@ -1,41 +1,0 @@
-package com.edurda77.filmlibrary.domain
-
-import androidx.annotation.WorkerThread
-import com.edurda77.filmlibrary.data.*
-
-interface TheMDBRepoUseCace {
-    @WorkerThread
-    @Throws(Throwable::class)
-    fun getReposForSearchMovieSync (userName: String, adultKey:Boolean): List<ResultSearchMovie>?
-    fun getReposForIDMovieSync (searcheMovie: ResultSearchMovie): Movie?
-    fun getReposForGenresSync():List<Genres>?
-    fun getReposForNowPlayingMovieSync (): List<ResultSearchMovie>?
-    fun getReposForPopularMovieSync (): List<ResultSearchMovie>?
-    fun getReposForTopRatedMovieSync (): List<ResultSearchMovie>?
-    fun getReposForUpcomingMovieSync (): List<ResultSearchMovie>?
-    fun getReposForSearchPeopleSync (name:String): List<ResultSearchedPeople>?
-    fun getReposForIdPeopleSync (searchedPeople:ResultSearchedPeople): People?
-
-
-    fun getReposForSearchMovieAsync (userName: String, adultKey:Boolean, onSuccess: (List<ResultSearchMovie>)->Unit,
-                                     OnError: (Throwable) ->Unit)
-    fun getReposForIDMovieAsync (searcheMovie: ResultSearchMovie,
-                                 onSuccess: (Movie)->Unit,
-                                 OnError: (Throwable) ->Unit)
-    fun getReposForGenresAsync(onSuccess: (List<Genres>)->Unit,
-                               OnError: (Throwable) ->Unit)
-    fun getReposForNowPlayingMovieAsync (onSuccess: (List<ResultSearchMovie>)->Unit,
-                                     OnError: (Throwable) ->Unit)
-    fun getReposForPopularMovieAsync (onSuccess: (List<ResultSearchMovie>)->Unit,
-                                     OnError: (Throwable) ->Unit)
-    fun getReposForTopRatedMovieAsync (onSuccess: (List<ResultSearchMovie>)->Unit,
-                                     OnError: (Throwable) ->Unit)
-    fun getReposForUpcomingMovieAsync (onSuccess: (List<ResultSearchMovie>)->Unit,
-                                     OnError: (Throwable) ->Unit)
-    fun getReposForSearchPeopleAsync (name: String, onSuccess: (List<ResultSearchedPeople>)->Unit,
-                                     OnError: (Throwable) ->Unit)
-    fun getReposForIdPeopleAsync (searchedPeople:ResultSearchedPeople,
-                                 onSuccess: (People)->Unit,
-                                 OnError: (Throwable) ->Unit)
-
-}
