+++ app/src/main/java/com/edurda77/filmlibrary/ui/SearchActivity.kt
@@ -13,13 +13,13 @@ import com.edurda77.filmlibrary.R
 import com.edurda77.filmlibrary.data.Movie
 import com.edurda77.filmlibrary.data.ResultSearchMovie
 import com.edurda77.filmlibrary.databinding.ActivitySearchBinding
-import com.edurda77.filmlibrary.domain.TheMDBRepoUseCace
+import com.edurda77.filmlibrary.domain.TheMDBRepoUseCaseSync
 
 private const val DEFAUL_KEY = "DEFAUL_KEY"
 class SearchActivity : AppCompatActivity() {
     private var toolbar: Toolbar? = null
     private lateinit var binding: ActivitySearchBinding
-    private val goSearchMovie: TheMDBRepoUseCace by lazy { app.theMDBRepoUseCace }
+    private val goSearchMovie: TheMDBRepoUseCaseSync by lazy { app.theMDBRepoUseCaceSync }
     private val resultSearch = emptyList<ResultSearchMovie>().toMutableList()
     private val prefernces: SharedPreferences by lazy {app.sharedPrefernces}
 
@@ -55,7 +55,7 @@ class SearchActivity : AppCompatActivity() {
     fun setOotRecycledView() {
 
         val recyclerView: RecyclerView = binding.itemSearchMovie
-        val goIDMovie: TheMDBRepoUseCace by lazy { app.theMDBRepoUseCace }
+        val goIDMovie: TheMDBRepoUseCaseSync by lazy { app.theMDBRepoUseCaceSync }
         recyclerView.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
 
         val stateClickListener: MovieAdapter.OnStateClickListener =
